fix tests

+++ inference-engine/tests/functional/inference_engine/transformations/softplus_decomposition_test.cpp
@@ -34,10 +34,10 @@ TEST(TransformationTests, SoftPlusDecomposition) {
     }
 
     {
-        auto input = std::make_shared<ngraph::opset4::Parameter>(ngraph::element::f64, ngraph::Shape{3, 1, 2});
+        auto input = std::make_shared<ngraph::opset4::Parameter>(ngraph::element::f32, ngraph::Shape{3, 1, 2});
         auto exp = std::make_shared<ngraph::opset4::Exp>(input);
         auto add = std::make_shared<ngraph::opset4::Add>(exp,
-            ngraph::opset4::Constant::create(ngraph::element::f64, ngraph::Shape{1}, {1}));
+            ngraph::opset4::Constant::create(ngraph::element::f32, ngraph::Shape{1}, {1.0}));
         auto log = std::make_shared<ngraph::opset4::Log>(add);
 
         f_ref = std::make_shared<ngraph::Function>(ngraph::NodeVector{log}, ngraph::ParameterVector{input});
